Add to_fv3config_diag_file_config method
