+++ workflows/prognostic_c48_run/setup.py
@@ -5,7 +5,7 @@
 
 from setuptools import setup, find_packages
 
-requirements = ["fv3fit", "vcm", "jsonschema>=3.2", "dacite >= 1.6"]
+requirements = ["fv3fit", "vcm", "jsonschema>=3.2", "dacite >= 1.6", "fv3config>=0.6.0"]
 
 test_requirements = ["pytest"]
